Add customization
    ^^^^^^^^^^^^^
just imagine that theres some glitter and fancy sparkles around that
i can't really add that in git

+++ linux/.bashrc
@@ -1,23 +1,36 @@
 # Check for interactive mode
 if [[ $- == *i* ]]; then
-	# NEW: SMART Welcome message
-	if ! (( $SHLVL > 1 )); then
-		clear
-		echo "Welcome, $USER!"
+	# Load user preferences
+	if [ -f ~/.bash_preferences ]; then source ~/.bash_preferences; fi
+
+
+	# Welcome message
+	if ! [ "$hide_welcome_message" = "1" ]; then
+		if ! (( $SHLVL > 1 )); then
+			clear
+			echo "Welcome, $USER!"
+		fi
+		date "+%A %d. %B %Y - %H:%M"
+		echo -en "bash $(if ! [ ${BASH_VERSINFO[4]} = "release" ];then echo "${BASH_VERSINFO[4]} ";fi )${BASH_VERSINFO[0]}.${BASH_VERSINFO[1]}\e[90m.${BASH_VERSINFO[2]}.${BASH_VERSINFO[3]}\e[0m"
+		if (( $SHLVL > 1 )); then echo -e " - nested level $(( $SHLVL - 1 ))"; else echo ""; fi
 	fi
-	date "+%A %d. %B %Y - %H:%M"
-	echo -en "bash $(if ! [ ${BASH_VERSINFO[4]} = "release" ];then echo "${BASH_VERSINFO[4]} ";fi )${BASH_VERSINFO[0]}.${BASH_VERSINFO[1]}\e[90m.${BASH_VERSINFO[2]}.${BASH_VERSINFO[3]}\e[0m"
-	if (( $SHLVL > 1 )); then echo -e " - nested level $(( $SHLVL - 1 ))"; else echo ""; fi
-	
+
+
+	# Command counter
 	_COMMANDS=-1
 	PROMPT_COMMAND="let _COMMANDS++"
+
 
 	# Set up prompt
 	_exitcode () {
 		if ! [ $_COMMANDS = 0 ]; then 
 			if [ $1 = 0 ]; then
+				if ! [ "$display_zero_exitcode" = "1" ]; then return 0; fi
 				echo -en "\e[90m"
+			else
+				echo -en "\e[1;91m"
 			fi
+
 			echo -en "Process exited with code $1\n\u200B"
 		fi
 	}
@@ -36,17 +49,19 @@ if [[ $- == *i* ]]; then
 \$(_pWd)\
 \$(_suffix)"
 
-fi
 
-# Git aliases
-alias yeet="git push"
-alias yoink="git pull"
-alias yikes="git reset --hard"
-yes () {
-	if git add $@; then
-		git commit
+	# Git aliases
+	_git_addandcommit () {
+		if git add $@; then
+			git commit
+		fi
+	}
+	if ! [ "$disable_git_aliases" = "1" ]; then
+		alias yeet="git push"
+		alias yoink="git pull"
+		alias yikes="git reset --hard"
+		alias yes=_git_addandcommit
+		alias yo="git status"
+		alias yaa="git push -f"
 	fi
-}
-alias yo="git status"
-alias yaa="git push -f"
-
+fi
